% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  dbhost = "localhost",
  dbport = 5432L,
  dbname = NULL,
  dbusername = NULL,
  dbpassword = NULL,
  credentials_path = NULL,
  credentials_pass = NULL,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{credentials_path}{Path to the SQLite database for shinymanager.}

\item{credentials_pass}{Passphrase for the SQLite database.}

\item{onStart}{A function that will be called before the app is actually run.}

\item{options}{Named options that should be passed to the `runApp` call.}

\item{enableBookmarking}{Can be one of "url", "server", or "disable".}

\item{uiPattern}{A regular expression used to determine which UI should be served.}

\item{...}{arguments to pass to golem_opts.}
}
\description{
Run the Shiny Application
}
